cria abstracao chamada ChartComponent para conter as informacoes base de um component cdo tipo chart
